+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2.2.2
+* Show repair cost in item tooltip when on the anvil screen
+* Rework enchantment hook due to incompatible mods
+
+## 2.2.1
+* Fix inventory buttons not working
+* Rework mixin for compatibility with enchantment table display #448
+* Remove custom name nbt from variant chests, barrels and crates
+* Added aerial affinity (F43nd1r)
+* Added extra recipes (Coranthes)
+
 ## 2.2.0
 * Fix Wandering Trader not spawning (F43nd1r)
 * Fix grindstone xp exploit (sven)